requirements.txt
asyncpg==0.23.0
discord.py==1.7.2
python-dotenv==0.17.1